refactor(client/terminal): extract mobileTouch.ts — isolate iOS-Safari volatility
Per /lowy audit: the proposed `@kolu/solid-xterm` framework
extraction *fails* the reuse test (one consumer; framework API
would be shaped around implementation, not stable contract). But
the 110 LOC of mobile-touch handling in `Terminal.tsx` is a real
encapsulation opportunity — it's the receptacle for an
independently-volatile axis (iOS Safari focus shuffling, xterm 6.x
touch-viewport gap, soft-keyboard summon heuristics).

Two helpers, separate concerns, same axis:
- `setupMobileTapToFocus(term)` — coerces `.xterm-screen` into
  contenteditable + bridges tap → term.focus() via a movement-
  thresholded pointerup. The dance exists because iOS Safari
  rejects soft-keyboard summoning when focus shuffles mid-gesture.
- `setupMobileTouchScroll(container, getTerm)` — bridges container
  touchmove → `terminal.scrollLines(...)` since xterm 6.0.0 ships
  typed `IViewport.handleTouchStart/Move` but no implementation,
  and the WebGL canvas eats touch events before they reach
  `.xterm-viewport`.

Both register listeners via `makeEventListener`, so they must run
inside a SolidJS owner — Terminal.tsx already calls them from
inside its `runWithOwner`-wrapped onMount tail.

Behavior preserving. Terminal.tsx 930→821 LOC.
Client unit tests: 172/172 pass.
