+++ docs/architecture-ralph-report.md
@@ -100,7 +100,8 @@ The repo has clearly internalized the extraction discipline:
 | Cycle | Target | Mutation | Δ violations | Δ LOC | Commit | Notes |
 | ----: | ------ | -------- | ------------ | ----- | ------ | ----- |
 |     0 | (baseline) | report scaffold | n/a | n/a | `afc9a35b` | establishes scorecard |
-|     1 | server import cycles | extract `surfaceCtx.ts` holder + pure `unwrapGit.ts` | −7 `noImportCycles` (7→0) | net +1 file, ~−30 LOC in `surface.ts` | (cycle 1) | All 7 server import cycles eliminated. Domain modules (`session.ts`, `activity.ts`, `terminalBackend/*`) now depend on `surfaceCtx.ts` only — a one-way arrow. `surface.ts` populates the holder via `setSurfaceCtx(...)` at module init. Two test files needed bootstrap (`metadata.test.ts` installs a no-op via the new `installNoopSurfaceCtxForTesting` helper; `session.test.ts` does a side-effect import of `surface.ts` since it verifies the real persistence path). 46/46 server unit tests pass; full repo unit suite green. Biome warnings 60→51. |
+|     1 | server import cycles | extract `surfaceCtx.ts` holder + pure `unwrapGit.ts` | −7 `noImportCycles` (7→0) | net +1 file, ~−30 LOC in `surface.ts` | `b7af5bdb` | All 7 server import cycles eliminated. Domain modules (`session.ts`, `activity.ts`, `terminalBackend/*`) now depend on `surfaceCtx.ts` only — a one-way arrow. `surface.ts` populates the holder via `setSurfaceCtx(...)` at module init. Two test files needed bootstrap (`metadata.test.ts` installs a no-op via the new `installNoopSurfaceCtxForTesting` helper; `session.test.ts` does a side-effect import of `surface.ts` since it verifies the real persistence path). 46/46 server unit tests pass; full repo unit suite green. Biome warnings 60→51. |
+|     2 | `Terminal.tsx` (930 LOC, mixed concerns) | extract `mobileTouch.ts` — `setupMobileTapToFocus(term)` + `setupMobileTouchScroll(container, getTerm)` | 0 lint Δ; isolates iOS-Safari platform volatility (Lowy Axis 8) | Terminal.tsx 930→821 (−109); `mobileTouch.ts` +135 | (cycle 2) | /lowy audit said the proposed `@kolu/solid-xterm` framework extraction *fails* the reuse test (one consumer; would shape API around implementation, not stable contract). What it *did* recommend was the mobile-touch state machine: 110 LOC of iOS-tap-vs-scroll heuristic + xterm-touchscroll bridge with a real platform-volatility axis. The receptacle is a single file; the axis (iOS Safari focus/touch behavior) now has a documented home for the next workaround. 172/172 client unit tests pass. |
 
 ---
